fix-to(impl): Support for module introduction
