+++ json-diff-common/src/main/java/me/codeleep/jsondiff/common/model/neat/JsonBuilder.java
@@ -1,0 +1,12 @@
+package me.codeleep.jsondiff.common.model.neat;
+
+/**
+ * @author: codeleep
+ * @createTime: 2024/04/11 下午7:08
+ * @description:
+ */
+public interface JsonBuilder {
+
+    JsonDiff builder(String json);
+
+}
