+++ json-diff-core/pom.xml
@@ -19,12 +19,6 @@
             <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
-        <dependency>
-            <groupId>cn.xiaoandcai</groupId>
-            <artifactId>json-diff-impl</artifactId>
-            <version>${project.version}</version>
-            <scope>compile</scope>
-        </dependency>
     </dependencies>
 
     <properties>
