+++ json-diff-core/src/main/java/me/codeleep/jsondiff/core/utils/JsonDiffBuilder.java
@@ -2,8 +2,9 @@ package me.codeleep.jsondiff.core.utils;
 
 import me.codeleep.jsondiff.common.model.neat.JsonDiff;
 import me.codeleep.jsondiff.core.config.JsonDiffOption;
-import me.codeleep.jsondiff.impl.FrameworkImplUtil;
-import me.codeleep.jsondiff.impl.ImplType;
+import me.codeleep.jsondiff.common.utils.ImplType;
+
+import static me.codeleep.jsondiff.common.utils.ImplType.*;
 
 /**
  * @author: codeleep
@@ -15,10 +16,10 @@ public class JsonDiffBuilder {
     public static JsonDiff buildObject(String jsonStr) {
         ImplType defaultJsonFramework = JsonDiffOption.getDefaultJsonFramework();
         switch (defaultJsonFramework) {
-            case FASTJSON: return FrameworkImplUtil.fastJson(jsonStr);
-            case FASTJSON2: return FrameworkImplUtil.fastJson2(jsonStr);
-            case GSON: return FrameworkImplUtil.gson(jsonStr);
-            case JACKSON: return FrameworkImplUtil.jackson(jsonStr);
+            case FASTJSON: return FASTJSON.getBuilder().builder(jsonStr);
+            case FASTJSON2: return FASTJSON2.getBuilder().builder(jsonStr);
+            case GSON: return GSON.getBuilder().builder(jsonStr);
+            case JACKSON: return JACKSON.getBuilder().builder(jsonStr);
         }
         return null;
     }
