+++ json-diff-impl/json-diff-impl-fastjson2/src/main/java/me/codeleep/jsondiff/impl/fastjson2/FastJson2Primitive.java
