+++ json-diff-test/pom.xml
@@ -20,13 +20,13 @@
     <dependencies>
         <dependency>
             <groupId>cn.xiaoandcai</groupId>
-            <artifactId>json-diff-common</artifactId>
+            <artifactId>json-diff</artifactId>
             <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
             <groupId>cn.xiaoandcai</groupId>
-            <artifactId>json-diff</artifactId>
+            <artifactId>json-diff-impl-gson</artifactId>
             <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
@@ -34,21 +34,6 @@
             <groupId>com.alibaba.fastjson2</groupId>
             <artifactId>fastjson2</artifactId>
             <version>2.0.7</version>
-        </dependency>
-        <dependency>
-            <groupId>com.alibaba</groupId>
-            <artifactId>fastjson</artifactId>
-            <version>1.2.74</version>
-        </dependency>
-        <dependency>
-            <groupId>com.google.code.gson</groupId>
-            <artifactId>gson</artifactId>
-            <version>2.9.1</version>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <version>2.13.0</version>
         </dependency>
         <dependency>
             <groupId>org.testng</groupId>
